refactor: implemente FSD and add property '.key'
