--- src/components/App.jsx
@@ -1,8 +1,0 @@
-import { Layout } from "./Layout.jsx";
-import { TabsRestaurants } from "./TabsRestaurants.jsx";
-
-export function App() {
-	return <Layout>
-		<TabsRestaurants />
-	</Layout>
-}
