--- src/components/Header.jsx
@@ -1,3 +1,0 @@
-export function Header() {
-	return <header className="header">HEADER</header>
-}
