+++ src/components/app/App.jsx
@@ -1,16 +1,8 @@
-import { headphones } from "../../constants/mock";
-import { Headphone } from "../headphone/Headphone";
-import { Layout } from "../layout/Layout";
+import { Layout } from "./Layout.jsx";
+import { TabsRestaurants } from "../widgets/TabsRestaurants.jsx";
 
-export const App = ({ title }) => {
-  return (
-    <div>
-      <Layout>
-        <h1>{title}</h1>
-        {headphones.map(({ name, reviews, codecs }) => (
-          <Headphone name={name} reviews={reviews} codecs={codecs} />
-        ))}
-      </Layout>
-    </div>
-  );
-};
+export function App() {
+	return <Layout>
+		<TabsRestaurants />
+	</Layout>
+}
