+++ src/components/widgets/Footer.jsx
@@ -2,5 +2,4 @@ export function Footer() {
 	return <footer className="footer">
 		FOOTER
 	</footer>
-
 }
