+++ src/components/widgets/Header.jsx
@@ -1,0 +1,5 @@
+export function Header() {
+	return <header className="header">
+		HEADER
+	</header>
+}
