+++ src/components/widgets/TabsRestaurants.jsx
@@ -1,9 +1,9 @@
-import { restaurants } from "../../materials/mock.js";
-import { Restaurant } from "./Restaurant.jsx";
+import { restaurants } from "../../../materials/mock.js";
+import { Restaurant } from "../entities/Restaurant.jsx";
 import { useState } from 'react'
 
 export function TabsRestaurants() {
-	const startRestaurantTab = restaurants.find((restaurantItem) => restaurantItem.name); // на случай если ресторан с индексом 0 будет без имени, т.е. невалидный. Иначе, если удалить имя у первого ресторана, то первый рендер будет баганый (будет постой таб).
+	const startRestaurantTab = restaurants.find((restaurantItem) => restaurantItem.name); // на случай если ресторан с индексом 0 будет без имени, т.е. невалидный. Иначе, если удалить имя у первого ресторана, то первый рендер будет баганый (будет пустой таб).
 	const startId = startRestaurantTab.id;
 	const [currId, setCurrId] = useState(startId);
 
@@ -13,8 +13,8 @@ export function TabsRestaurants() {
 				{
 					restaurants.map((restaurant) => !!restaurant.name && (
 						<div
-							onClick={event => setCurrId(event.target.dataset.id)}
-							data-id={restaurant.id}
+							key={restaurant.id}
+							onClick={() => setCurrId(restaurant.id)}
 							className="restaurants-tabs__title"
 						>{restaurant.name}</div>
 					))
