+++ src/main.jsx
@@ -1,5 +1,5 @@
 import { createRoot } from "react-dom/client";
-import { App } from "/src/components/App.jsx";
+import { App } from "/src/components/app/App.jsx";
 
 const rootElement = document.getElementById("root");
